Update package-lock.json
I was not able to reproduce the error on my personal computer. I recall
it was something like `addTransaction is not a function`. The error was
located at `src/components/AddTransaction.js` line 19.

```
    addTransaction(newTransaction);
```

The only thing I can think of is that updating the package lock file may
have fixed the issue, but I doubt this is the case. The package lock
file specifies project dependencies while `addTransaction()` is a
function defined in the application.